requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.7
django-debug-toolbar==1.9.1
django-environ==0.4.5
django-filter==1.1.0
django-toolbelt==0.0.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
flake8==3.5.0
future==0.16.0
gunicorn==19.9.0
isort==4.3.4
mccabe==0.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
